requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9

# HTTP Clients
httpx==0.26.0
requests==2.31.0

# LLM Clients
openai==1.10.0
google-generativeai==0.3.2
anthropic==0.8.1

# Telegram
python-telegram-bot==20.7

# Utils
python-dotenv==1.0.0
python-multipart==0.0.22

# Logging e Monitoring
structlog==24.1.0
sentry-sdk==1.39.2

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx-mock==0.15.0

# Video Processing
moviepy==1.0.3

# Storage
boto3==1.34.26

# Scheduling
schedule==1.2.0

# Data Validation
validators==0.22.0

# Date/Time
pytz==2023.3
python-dateutil==2.8.2
